refactor duplicate code of app indexing calls
