+++ app/src/main/java/org/gdg/frisbee/android/common/GdgActivity.java
@@ -25,10 +25,14 @@ import android.os.Bundle;
 import android.os.Handler;
 import android.support.v7.widget.Toolbar;
 
+import com.google.android.gms.appindexing.Action;
 import com.google.android.gms.appindexing.AppIndex;
 import com.google.android.gms.appstate.AppStateManager;
 import com.google.android.gms.common.ConnectionResult;
 import com.google.android.gms.common.api.GoogleApiClient;
+import com.google.android.gms.common.api.PendingResult;
+import com.google.android.gms.common.api.ResultCallback;
+import com.google.android.gms.common.api.Status;
 import com.google.android.gms.games.Games;
 import com.google.android.gms.plus.Plus;
 
@@ -42,6 +46,7 @@ import org.gdg.frisbee.android.utils.RecentTasksStyler;
 import org.gdg.frisbee.android.utils.Utils;
 
 import butterknife.ButterKnife;
+import timber.log.Timber;
 
 public abstract class GdgActivity extends TrackableActivity implements
         GoogleApiClient.ConnectionCallbacks, GoogleApiClient.OnConnectionFailedListener {
@@ -276,5 +281,28 @@ public abstract class GdgActivity extends TrackableActivity implements
 
     public void setToolbarTitle(final String title) {
         getActionBarToolbar().setTitle(title);
+    }
+
+    protected void recordEndPageView(Action viewAction) {
+        PendingResult<Status> result = AppIndex.AppIndexApi.end(getGoogleApiClient(), viewAction);
+        result.setResultCallback(appIndexApiCallback("end " + viewAction));
+    }
+
+    protected void recordStartPageView(Action viewAction) {
+        PendingResult<Status> result = AppIndex.AppIndexApi.start(getGoogleApiClient(), viewAction);
+        result.setResultCallback(appIndexApiCallback("start " + viewAction));
+    }
+
+    protected ResultCallback<Status> appIndexApiCallback(final String label) {
+        return new ResultCallback<Status>() {
+            @Override
+            public void onResult(Status status) {
+                if (status.isSuccess()) {
+                    Timber.d("App Indexing API: Recorded event %s view successfully.", label);
+                } else {
+                    Timber.e("App Indexing API: There was an error recording the event view. Status = %s", status.toString());
+                }
+            }
+        };
     }
 }
